# sunset-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sunset Rollercoaster API",
    "description": "Read-only REST API for Taipower electricity data, Taiwan and international fuel prices, invoice draws, and reservoir records.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/taipower/area-loads": {
      "get": {
        "tags": [
          "taipower"
        ],
        "operationId": "list_area_loads",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "description": "Exact local date in Asia/Taipei (YYYY-MM-DD).",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "start",
            "in": "query",
            "description": "Inclusive UTC timestamp lower bound (RFC3339). Ignored when `date` is set.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end",
            "in": "query",
            "description": "Inclusive UTC timestamp upper bound (RFC3339). Ignored when `date` is set.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 500,
              "maximum": 5000,
              "minimum": 1
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List ten-minute regional load curve records; power values are MW",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaipowerAreaLoad"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/taipower/area-loads/latest": {
      "get": {
        "tags": [
          "taipower"
        ],
        "operationId": "latest_area_load",
        "responses": {
          "200": {
            "description": "Latest ten-minute regional load record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaipowerAreaLoad"
                }
              }
            }
          },
          "404": {
            "description": "No data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/taipower/area-snapshots": {
      "get": {
        "tags": [
          "taipower"
        ],
        "operationId": "list_area_snapshots",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "description": "Exact local date in Asia/Taipei (YYYY-MM-DD).",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "start",
            "in": "query",
            "description": "Inclusive UTC timestamp lower bound (RFC3339). Ignored when `date` is set.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end",
            "in": "query",
            "description": "Inclusive UTC timestamp upper bound (RFC3339). Ignored when `date` is set.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 500,
              "maximum": 5000,
              "minimum": 1
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List regional generation and load snapshots; power values are MW",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaipowerAreaSnapshot"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/taipower/area-snapshots/latest": {
      "get": {
        "tags": [
          "taipower"
        ],
        "operationId": "latest_area_snapshot",
        "responses": {
          "200": {
            "description": "Latest regional generation and load snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaipowerAreaSnapshot"
                }
              }
            }
          },
          "404": {
            "description": "No data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/taipower/fuel-mix": {
      "get": {
        "tags": [
          "taipower"
        ],
        "operationId": "list_fuel_mix",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "description": "Exact local date in Asia/Taipei (YYYY-MM-DD).",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "start",
            "in": "query",
            "description": "Inclusive UTC timestamp lower bound (RFC3339). Ignored when `date` is set.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end",
            "in": "query",
            "description": "Inclusive UTC timestamp upper bound (RFC3339). Ignored when `date` is set.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 500,
              "maximum": 5000,
              "minimum": 1
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List ten-minute fuel-mix curve records; power values are MW",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaipowerFuelMix"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/taipower/fuel-mix/latest": {
      "get": {
        "tags": [
          "taipower"
        ],
        "operationId": "latest_fuel_mix",
        "responses": {
          "200": {
            "description": "Latest ten-minute fuel-mix record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaipowerFuelMix"
                }
              }
            }
          },
          "404": {
            "description": "No data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/taipower/generators": {
      "get": {
        "tags": [
          "taipower"
        ],
        "operationId": "list_generators",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "description": "Exact snapshot date in Asia/Taipei (YYYY-MM-DD).",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "category_code",
            "in": "query",
            "description": "Exact source category code, for example `lng`, `coal`, or `solar`.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unit_name",
            "in": "query",
            "description": "Exact generator or subtotal name.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "is_summary",
            "in": "query",
            "description": "Filter regular unit rows (`false`) or category subtotal rows (`true`).",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 500,
              "maximum": 5000,
              "minimum": 1
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List generator output snapshot rows; power values are MW",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaipowerGenerator"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/taipower/operating-reserves": {
      "get": {
        "tags": [
          "taipower"
        ],
        "operationId": "list_operating_reserves",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "start",
            "in": "query",
            "description": "Inclusive date lower bound. Ignored when `date` is set.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end",
            "in": "query",
            "description": "Inclusive date upper bound. Ignored when `date` is set.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 500,
              "maximum": 5000,
              "minimum": 1
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List daily peak load and operating reserve records; power values are MW",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaipowerOperatingReserve"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/taipower/operating-reserves/latest": {
      "get": {
        "tags": [
          "taipower"
        ],
        "operationId": "latest_operating_reserve",
        "responses": {
          "200": {
            "description": "Latest operating reserve record (normally today's forecast)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaipowerOperatingReserve"
                }
              }
            }
          },
          "404": {
            "description": "No data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/taipower/power-snapshots/latest": {
      "get": {
        "tags": [
          "taipower"
        ],
        "operationId": "latest_power_snapshot",
        "responses": {
          "200": {
            "description": "Latest Taipower current-load and peak forecast snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaipowerPowerSnapshot"
                }
              }
            }
          },
          "404": {
            "description": "No data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "TaipowerAreaLoad": {
        "type": "object",
        "required": [
          "id",
          "observed_at",
          "north_load_mw",
          "central_load_mw",
          "south_load_mw",
          "east_load_mw",
          "total_load_mw"
        ],
        "properties": {
          "central_load_mw": {
            "type": "number",
            "format": "double"
          },
          "east_load_mw": {
            "type": "number",
            "format": "double"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "north_load_mw": {
            "type": "number",
            "format": "double"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time"
          },
          "south_load_mw": {
            "type": "number",
            "format": "double"
          },
          "total_load_mw": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
    }
  },
...
}